MAINT: Move add_2d_latlon_axis to util.transforms
