+++ pyart/util/transforms.py
@@ -14,6 +14,7 @@ Cartesian (x, y, z), cartographic (latitude, longitude, altitude) and antenna
     antenna_to_cartesian_track_relative
     antenna_to_cartesian_earth_relative
     antenna_to_cartesian_aircraft_relative
+    add_2d_latlon_axis
     _interpolate_axes_edges
     _interpolate_azimuth_edges
     _interpolate_elevation_edges
@@ -305,3 +306,103 @@ def antenna_to_cartesian_aircraft_relative(rng, rot, tilt):
     y = r * np.sin(tilt)
     z = r * np.cos(rot) * np.cos(tilt)
     return x, y, z
+
+
+def add_2d_latlon_axis(grid, **kwargs):
+    """
+    Add the latitude and longitude for grid points in the y, x plane.
+
+    Adds a **latitude** and **longitude** dictionary to the axes attribute
+    of a provided grid.  Addition is done in-place, nothing is returned from
+    this function.  These dictionaries contain 2D arrays which specify the
+    latitude and longitude of every point in the y, x plane.
+
+    If available, the conversion is done using basemap.pyproj, extra arguments
+    are passed to pyproj.Proj. If not available, an internal spherical
+    azimuthal equidistant transformation is is used.
+
+    Parameters
+    ----------
+    grid: grid object
+        Cartesian grid object containing the 1d axes "x_disp", "y_disp" and
+        scalar axes 'lat', 'lon'.
+    kwargs: Pyproj options
+        Options to be passed to Proj. If projection is not specified here it
+        uses proj='aeqd' (azimuthal equidistant)
+
+    Notes
+    -----
+    If Basemap is not available, calculation of the latitude, longitude is
+    done using a azimuthal equidistant projection projection [1].
+    It uses the mean radius of earth (6371 km)
+
+    .. math::
+
+        c = \\sqrt(x^2 + y^2)/R
+
+        azi = \\arctan2(y,x) \\text{  # from east to north}
+
+        lat = \\arcsin(\\cos(c)*\\sin(lat0)+\\sin(azi)*\\sin(c)*\\cos(lat0))
+
+        lon = \\arctan2(\\cos(azi)*\\sin(c),\\cos(c)*\\cos(lat0)-
+                        \\sin(azi)*\\sin(c)*\\sin(lat0)) + lon0
+
+    Where x, y are the Cartesian position from the center of projection;
+    lat, lon the corresponding latitude and longitude; lat0, lon0 the latitude
+    and longitude of the center of the projection; R the mean radius of the
+    earth (6371 km)
+
+    References
+    ----------
+    .. [1] Snyder, J. P. Map Projections--A Working Manual. U. S. Geological
+        Survey Professional Paper 1395, 1987, pp. 191-202.
+
+    """
+    try:
+        from mpl_toolkits.basemap import pyproj
+        if 'proj' not in kwargs:
+            kwargs['proj'] = 'aeqd'
+        x, y = np.meshgrid(
+            grid.axes["x_disp"]['data'], grid.axes["y_disp"]['data'])
+        b = pyproj.Proj(lat_0=grid.axes["lat"]['data'][0],
+                        lon_0=grid.axes["lon"]['data'][0], **kwargs)
+        lon, lat = b(x, y, inverse=True)
+    except ImportError:
+        import warnings
+        warnings.warn('No basemap found, using internal implementation '
+                      'for converting azimuthal equidistant to latlon')
+        # azimutal equidistant projetion to latlon
+        R = 6371.0 * 1000.0     # radius of earth in meters.
+
+        x, y = np.meshgrid(grid.axes["x_disp"]['data'],
+                           grid.axes["y_disp"]['data'])
+
+        c = np.sqrt(x*x + y*y) / R
+        phi_0 = grid.axes["lat"]['data'] * np.pi / 180
+        azi = np.arctan2(y, x)  # from east to north
+
+        lat = np.arcsin(np.cos(c) * np.sin(phi_0) +
+                        np.sin(azi) * np.sin(c) * np.cos(phi_0)) * 180 / np.pi
+        lon = (np.arctan2(np.cos(azi) * np.sin(c), np.cos(c) * np.cos(phi_0) -
+                          np.sin(azi) * np.sin(c) * np.sin(phi_0)) * 180 /
+               np.pi + grid.axes["lon"]['data'])
+        lon = np.fmod(lon + 180, 360) - 180
+
+    lat_axis = {
+        'data':  lat,
+        'long_name': 'Latitude for points in Cartesian system',
+        'axis': 'YX',
+        'units': 'degree_N',
+        'standard_name': 'latitude',
+    }
+
+    lon_axis = {
+        'data': lon,
+        'long_name': 'Longitude for points in Cartesian system',
+        'axis': 'YX',
+        'units': 'degree_E',
+        'standard_name': 'longitude',
+    }
+
+    grid.axes["latitude"] = lat_axis
+    grid.axes["longitude"] = lon_axis
